Layer - Network_Manager - Connection
The Network will be constructed by the HTM Theory of Numenta. There for it needs Layers, Connections and a Network_Manager.
The Layer is a group of Neurons that will process together. A Connection connects two Neurons in one Way ( from input to output  ) and the Network_Manager generates the Network and read the output and feed in the input form an environment .