Just changes in the Server/ directory Removed: NeuronalNetwork.py Connection.py Added: Layer.py README.md
In this Project I fix on the HTM Theory by numenta.org / numenta.com
I believe that they take the best and most robust way toward strong
AI.
In Layer.py will be an implementation of the executen group of
a neuron in the HTM, the Layer.
It implements two classes: Neuron and Layer.
Neuron has a ready init but nothing else and the Layer is TODO.
For a better and more transperant code I added the README.md.
It will be developed with each change.